requirements.txt
Flask==1.0.2
Flask-WTF==0.14.2
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
requests==2.21.0
gunicorn
